STATE PATTERN A lot of objects have states. For example, a radio. A radio have basically two states: on and off. Can we represent this in object oriented programming? Well, yes:
This is just a simple example, but, when we talk about an object that
have a lot of states, this control can help a lot. You can define rules
to create final states and states that requires an previous state to be
executed. For example, in this radio scenario, you can only turn a radio
off if the radio is on, otherwise an exception will be thrown. Basically
you can do whatever the business tell you to do.